feat: Merge v2.0.0 milestone branch in next release.

+++ classes/class-uagb-admin-helper.php
@@ -277,16 +277,24 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 
 			$wp_upload_dir = UAGB_Helper::get_uag_upload_dir_path();
 			$combined_path = $wp_upload_dir . 'custom-style-blocks.css';
-			wp_delete_file( $combined_path );
+
+			if ( file_exists( $combined_path ) ) {
+				wp_delete_file( $combined_path );
+			}
 
 			$style = '';
 
 			$wp_filesystem = uagb_filesystem();
 
 			foreach ( $combined as $key => $c_block ) {
-				$style .= $wp_filesystem->get_contents( UAGB_DIR . 'assets/css/blocks/' . $c_block . '.css' );
 
+				$style_file = UAGB_DIR . 'assets/css/blocks/' . $c_block . '.css';
+
+				if ( file_exists( $style_file ) ) {
+					$style .= $wp_filesystem->get_contents( $style_file );
+				}
 			}
+
 			$wp_filesystem->put_contents( $combined_path, $style, FS_CHMOD_FILE );
 		}
 
@@ -301,7 +309,7 @@ if ( ! class_exists( 'UAGB_Admin_Helper' ) ) {
 
 			$rollback_versions = get_transient( 'uag_rollback_versions_' . UAGB_VER );
 
-			if ( ! $rollback_versions || empty( $rollback_versions ) ) {
+			if ( empty( $rollback_versions ) ) {
 
 				$max_versions = 10;
